+++ apps/backoffice-v2/src/pages/Entity/components/CallToAction/hooks/useCallToActionLogic/useCallToActionLogic.tsx
@@ -1,22 +1,41 @@
+import { CommonWorkflowEvent } from '@ballerine/common';
+import { useCallback, useEffect, useState } from 'react';
+import toast from 'react-hot-toast';
 import { useParams } from 'react-router-dom';
 import { useFilterId } from '../../../../../../common/hooks/useFilterId/useFilterId';
-import { useWorkflowQuery } from '../../../../../../domains/workflows/hooks/queries/useWorkflowQuery/useWorkflowQuery';
 import { useAuthenticatedUserQuery } from '../../../../../../domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
-import { useCaseState } from '../../../Case/hooks/useCaseState/useCaseState';
-import toast from 'react-hot-toast';
-import { useCallback, useEffect, useState } from 'react';
 import { useApproveTaskByIdMutation } from '../../../../../../domains/entities/hooks/mutations/useApproveTaskByIdMutation/useApproveTaskByIdMutation';
 import { useRejectTaskByIdMutation } from '../../../../../../domains/entities/hooks/mutations/useRejectTaskByIdMutation/useRejectTaskByIdMutation';
 import { useRevisionTaskByIdMutation } from '../../../../../../domains/entities/hooks/mutations/useRevisionTaskByIdMutation/useRevisionTaskByIdMutation';
 import { TWorkflowById } from '../../../../../../domains/workflows/fetchers';
-import { CommonWorkflowEvent } from '@ballerine/common';
+import { useWorkflowQuery } from '../../../../../../domains/workflows/hooks/queries/useWorkflowQuery/useWorkflowQuery';
+import { useCaseState } from '../../../Case/hooks/useCaseState/useCaseState';
+import { ICallToActionDocumentSelection } from '@/pages/Entity/components/CallToAction/interfaces';
+import { useDocumentSelection } from '@/pages/Entity/components/CallToAction/hooks/useDocumentSelection';
+
+export interface UseCallToActionLogicParams {
+  contextUpdateMethod?: 'base' | 'director';
+  revisionReasons?: string[];
+  rejectionReasons?: string[];
+  documentSelection?: ICallToActionDocumentSelection;
+  onReuploadReset?: () => void;
+  onDialogClose?: () => void;
+}
 
 const getPostUpdateEventNameEvent = (workflow: TWorkflowById) => {
   if (!workflow?.workflowDefinition?.config?.workflowLevelResolution) {
     return CommonWorkflowEvent.TASK_REVIEWED;
   }
 };
-export const useCallToActionLogic = () => {
+export const useCallToActionLogic = (params: UseCallToActionLogicParams) => {
+  const {
+    contextUpdateMethod = 'base',
+    rejectionReasons,
+    revisionReasons,
+    documentSelection,
+    onReuploadReset,
+    onDialogClose,
+  } = params;
   const { entityId } = useParams();
   const filterId = useFilterId();
   const { data: workflow } = useWorkflowQuery({ workflowId: entityId, filterId });
@@ -31,14 +50,6 @@ export const useCallToActionLogic = () => {
   const { mutate: mutateRevisionTaskById, isLoading: isLoadingRevisionTaskById } =
     useRevisionTaskByIdMutation(workflow?.id, postUpdateEventName);
 
-  const revisionReasons =
-    workflow?.workflowDefinition?.contextSchema?.schema?.properties?.documents?.items?.properties?.decision?.properties?.revisionReason?.anyOf?.find(
-      ({ enum: enum_ }) => !!enum_,
-    )?.enum;
-  const rejectionReasons =
-    workflow?.workflowDefinition?.contextSchema?.schema?.properties?.documents?.items?.properties?.decision?.properties?.rejectionReason?.anyOf?.find(
-      ({ enum: enum_ }) => !!enum_,
-    )?.enum;
   const isLoadingTaskDecisionById =
     isLoadingApproveTaskById || isLoadingRejectTaskById || isLoadingRevisionTaskById;
 
@@ -58,6 +69,11 @@ export const useCallToActionLogic = () => {
   const noReasons = !reasons?.length;
   const [reason, setReason] = useState(reasons?.[0] ?? '');
   const [comment, setComment] = useState('');
+
+  const resetReasonAndComment = useCallback(() => {
+    setReason('');
+    setComment('');
+  }, []);
 
   const onReasonChange = useCallback((value: string) => setReason(value), [setReason]);
   const onActionChange = useCallback((value: typeof action) => setAction(value), [setAction]);
@@ -86,6 +102,7 @@ export const useCallToActionLogic = () => {
         if (payload?.decision === 'approve') {
           return mutateApproveTaskById({
             documentId: payload?.id,
+            contextUpdateMethod,
           });
         }
 
@@ -107,12 +124,13 @@ export const useCallToActionLogic = () => {
             documentId: payload?.id,
             reason: payload?.reason,
             decision: payload?.decision,
+            contextUpdateMethod,
           });
         }
 
         toast.error('Invalid decision');
       },
-    [mutateApproveTaskById, mutateRejectTaskById, mutateRevisionTaskById],
+    [contextUpdateMethod, mutateApproveTaskById, mutateRejectTaskById, mutateRevisionTaskById],
   );
   const workflowLevelResolution =
     workflow?.workflowDefinition?.config?.workflowLevelResolution ??
@@ -122,6 +140,18 @@ export const useCallToActionLogic = () => {
     setReason(reasons?.[0] ?? '');
   }, [action, reasons]);
 
+  const isReuploadResetable = Boolean(onReuploadReset);
+
+  const documentPickerProps = useDocumentSelection(documentSelection, resetReasonAndComment);
+
+  const handleDialogClose = useCallback(
+    (open: boolean) => {
+      // Calling callback only when dialog is closed.
+      if (open === false) onDialogClose && onDialogClose();
+    },
+    [onDialogClose],
+  );
+
   return {
     isLoadingTaskDecisionById,
     caseState,
@@ -130,11 +160,14 @@ export const useCallToActionLogic = () => {
     reasons,
     reason,
     comment,
+    documentPickerProps,
     onReasonChange,
     onActionChange,
     onCommentChange,
+    handleDialogClose,
     noReasons,
     onMutateTaskDecisionById,
     workflowLevelResolution,
+    isReuploadResetable,
   };
 };