+++ apps/kyb-app/src/common/components/molecules/ProgressBar/ProgressBar.tsx
@@ -2,14 +2,16 @@ import styles from './ProgressBar.module.css';
 import { Chip } from '@/common/components/atoms/Chip';
 import { LoadingSpinner } from '@/common/components/atoms/LoadingSpinner';
 import { Check } from 'lucide-react';
-import clsx from 'clsx';
+import { ctw } from '@ballerine/ui';
 import { useDynamicUIContext } from '@/components/organisms/DynamicUI/hooks/useDynamicUIContext';
+import { useTranslation } from 'react-i18next';
 
 interface Props {
   className?: string;
 }
 
 export const ProgressBar = ({ className }: Props) => {
+  const { t } = useTranslation();
   const { state } = useDynamicUIContext();
   const { isLoading } = state;
 
@@ -20,7 +22,7 @@ export const ProgressBar = ({ className }: Props) => {
           <LoadingSpinner size="14" />
         ) : (
           <div
-            className={clsx(
+            className={ctw(
               'flex h-3 w-3 items-center justify-center rounded-full bg-[#00BD59]',
               styles.bounceAnimation,
             )}
@@ -31,7 +33,7 @@ export const ProgressBar = ({ className }: Props) => {
       }
       className={className}
       variant={isLoading ? 'primary' : 'success'}
-      text={isLoading ? 'Saving' : 'Progress saved'}
+      text={isLoading ? t('saving') : t('progressSaved')}
     />
   );
 };