+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/components/FileUploaderField/FileUploaderField.tsx
@@ -1,0 +1,65 @@
+import { useFileAssigner } from '@/components/organisms/UIRenderer/elements/JSONForm/components/FileUploaderField/hooks/useFileAssigner';
+import { useFileRepository } from '@/components/organisms/UIRenderer/elements/JSONForm/components/FileUploaderField/hooks/useFileRepository';
+import { useFileUploading } from '@/components/organisms/UIRenderer/elements/JSONForm/components/FileUploaderField/hooks/useFileUploading';
+import { DocumentUploadFieldProps } from '@/components/organisms/UIRenderer/elements/JSONForm/components/FileUploaderField/types';
+import { Input } from '@ballerine/ui';
+import { forwardRef, useCallback, useRef } from 'react';
+
+export const FileUploaderField = forwardRef(
+  ({
+    onChange,
+    onBlur,
+    uploadFile: _uploadFile,
+    fileRepository: fileStorage,
+    fileId,
+    isLoading,
+    disabled,
+    acceptFileFormats,
+    placeholder,
+  }: DocumentUploadFieldProps) => {
+    const { isUploading, uploadFile } = useFileUploading(_uploadFile);
+    const { file: registeredFile, registerFile } = useFileRepository(
+      fileStorage,
+      fileId || undefined,
+    );
+    const inputRef = useRef<HTMLInputElement>(null);
+    //@ts-ignore
+    useFileAssigner(inputRef, registeredFile);
+
+    const handleChange = useCallback(
+      async (event: React.ChangeEvent<HTMLInputElement>) => {
+        const file = event.target.files?.[0];
+
+        if (!file) return;
+
+        const uploadResult = await uploadFile(file);
+
+        registerFile(file, uploadResult.fileId);
+
+        const registeredFile = fileStorage.getFileById(uploadResult.fileId);
+
+        if (!registeredFile) {
+          throw new Error(
+            `Failed to register file with the id of ${uploadResult.fileId} in the file repository`,
+          );
+        }
+
+        onChange(uploadResult.fileId);
+      },
+      [uploadFile, registerFile, fileStorage, onChange],
+    );
+
+    return (
+      <Input
+        data-testid="file-uploader-field"
+        type="file"
+        placeholder={placeholder}
+        accept={acceptFileFormats}
+        disabled={disabled || isLoading || isUploading}
+        onChange={handleChange}
+        onBlur={onBlur}
+        ref={inputRef}
+      />
+    );
+  },
+);