+++ services/workflows-service/scripts/workflows/dynamic-ui-workflow.ts
@@ -74,8 +74,7 @@ export const dynamicUiWorkflowDefinition = {
             },
             kycAndVendorDone,
           ],
-          reject: 'rejected',
-          VENDOR_FAILED: 'rejected',
+          VENDOR_FAILED: 'failed',
         },
       },
       pending_kyc_response_to_finish: {
@@ -165,10 +164,10 @@ export const dynamicUiWorkflowDefinition = {
               companyName: entity.data.companyName,
               customerName: metadata.customerName,
               firstName: entity.data.additionalInfo.mainRepresentative.firstName,
-              collectionFlowUrl: join('',['{secret.COLLECTION_FLOW_URL}','/?token=',metadata.token]),
+              collectionFlowUrl: join('',['{secret.COLLECTION_FLOW_URL}','/?token=',metadata.token,'&lng=cn']),
               from: 'no-reply@ballerine.com',
               receivers: [entity.data.additionalInfo.mainRepresentative.email],
-              templateId: 'd-8949519316074e03909042cfc5eb4f02',
+              templateId: 'd-add814d1a97e4ea8b6db47c8f80de3cd',
               adapter: '{secret.MAIL_ADAPTER}'
               }`, // jmespath
             },
@@ -367,6 +366,7 @@ export const dynamicUiWorkflowDefinition = {
         deliverEvent: 'KYC_REVISION',
       },
     ],
+    workflowLevelResolution: true,
   },
   contextSchema: {
     type: 'json-schema',