+++ services/workflows-service/src/collection-flow/dto/get-flow-configuration-input.dto.ts
@@ -1,0 +1,6 @@
+import { IsString } from 'class-validator';
+
+export class GetFlowConfigurationInputDto {
+  @IsString()
+  language!: string;
+}