+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ballerine/backoffice-v2
 
+## 0.5.32
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/common@0.7.30
+  - @ballerine/ui@0.3.18
+  - @ballerine/workflow-browser-sdk@0.5.29
+  - @ballerine/workflow-node-sdk@0.5.29
+
 ## 0.5.31
 
 ### Patch Changes
@@ -54,6 +65,8 @@
 
 ### Patch Changes
 
+- @ballerine/workflow-browser-sdk@0.5.23
+- @ballerine/workflow-node-sdk@0.5.23
 - Updated dependencies
   - @ballerine/ui@0.3.16
 - Updated dependencies [d5cca1e8]
@@ -67,6 +80,7 @@
 ### Patch Changes
 
 - Updated dependencies [8c888cec]
+- Updated dependencies
   - @ballerine/workflow-browser-sdk@0.5.22
   - @ballerine/workflow-node-sdk@0.5.22
   - @ballerine/common@0.7.23
@@ -77,6 +91,7 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @ballerine/common@0.7.22
   - @ballerine/workflow-browser-sdk@0.5.21
   - @ballerine/workflow-node-sdk@0.5.21