+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,16 @@
 # kyb-app
 
+## 0.1.26
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/blocks@0.1.22
+  - @ballerine/common@0.7.30
+  - @ballerine/ui@0.3.18
+  - @ballerine/workflow-browser-sdk@0.5.29
+
 ## 0.1.25
 
 ### Patch Changes
@@ -49,6 +60,7 @@
 
 ### Patch Changes
 
+- @ballerine/workflow-browser-sdk@0.5.23
 - Updated dependencies
   - @ballerine/ui@0.3.16
 - Updated dependencies [d5cca1e8]
@@ -61,6 +73,7 @@
 ### Patch Changes
 
 - Updated dependencies [8c888cec]
+- Updated dependencies
   - @ballerine/workflow-browser-sdk@0.5.22
   - @ballerine/blocks@0.1.21
   - @ballerine/common@0.7.23
@@ -84,6 +97,8 @@
 
 ### Patch Changes
 
+- @ballerine/blocks@0.1.20
+- @ballerine/workflow-browser-sdk@0.5.20
 - Change document schemas
 - Updated dependencies
   - @ballerine/blocks@0.1.20