+++ apps/workflows-dashboard/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/workflows-dashboard
 
+## 0.1.10
+
+### Patch Changes
+
+- Version bump
+
 ## 0.1.9
 
 ### Patch Changes