+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflow-core
 
+## 0.5.29
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/common@0.7.30
+
 ## 0.5.28
 
 ### Patch Changes
@@ -44,6 +52,7 @@
 
 ### Patch Changes
 
+- Added TransformerPlugin
 - d5cca1e8: added defintion validation
 - f4b5c0df: added defintion validation
 - Updated dependencies [d5cca1e8]