+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ballerine/workflows-service
 
+## 0.5.29
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/common@0.7.30
+  - @ballerine/workflow-core@0.5.29
+  - @ballerine/workflow-node-sdk@0.5.29
+
 ## 0.5.28
 
 ### Patch Changes
@@ -26,6 +36,8 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @ballerine/workflow-core@0.5.23
+  - @ballerine/workflow-node-sdk@0.5.23
 - Updated dependencies [59afd0b4]
   - @ballerine/common@0.7.27
   - @ballerine/workflow-core@0.5.26